requirements.txt
beautifulsoup4==4.11.1
httpx==0.27.2
Jinja2==3.1.4
python_dateutil==2.8.2
PyYAML==6.0.1
Requests==2.32.3
lzma
zipfile